requirements.txt
# ============================================================
# GraDeT-HTR requirements
# Install PyTorch FIRST via the official selector:
#   https://pytorch.org/get-started/locally/
# Example for CUDA 12.1:
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121
# Then install this file:
#   pip install -r requirements.txt
# ============================================================

# --- BUET Bangla normalizer (must be first) ---
git+https://github.com/csebuetnlp/normalizer@d80c3c4#egg=normalizer

# --- Core ML ---
# torch and torchvision are installed separately (see above)
transformers==4.47.0
tokenizers==0.21.0
huggingface-hub==0.26.5
safetensors==0.4.5

# --- Data / numerics ---
numpy==2.0.2
pandas==2.2.3
scikit-learn==1.6.0
scipy==1.13.1

# --- Image processing ---
pillow==11.0.0
opencv-python-headless==4.11.0.86
albumentations==2.0.5

# --- Trie tokenizer ---
marisa-trie==1.2.1

# --- Training utilities ---
tqdm==4.67.1
psutil==7.0.0
py-cpuinfo==9.0.0
PyYAML==6.0.2

# --- Text utilities ---
regex==2024.11.6
ftfy==6.0.3
emoji==1.4.2

# --- Segmentation (inference pipeline only) ---
ultralytics==8.3.151

# --- PDF handling (inference pipeline only) ---
pdf2image==1.17.0

# --- Plotting / reporting ---
matplotlib==3.9.4
seaborn==0.13.2

# --- Misc ---
filelock==3.16.1
packaging==24.2
requests==2.32.3
sympy==1.13.1
networkx==3.2.1
typing_extensions==4.12.2
six==1.17.0
python-dateutil==2.9.0.post0
pytz==2024.2
tzdata==2024.2
PyYAML==6.0.2
rich==14.0.0
